pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finance-alchemy"
version = "0.1.0"
description = "A股智能量化分析系统 — 基于 CrewAI 多智能体框架与 DDD 架构"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["A-share", "stock", "crewai", "quantitative", "DDD", "china", "finance"]

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.17.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "akshare>=1.14.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.9.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "langchain-openai>=0.1.0",
    "litellm>=1.0.0",
    "gradio>=4.0.0",
    "cachetools>=5.3.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "joblib",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "types-requests>=2.32.0",
]

[project.scripts]
finance-alchemy = "src.interface.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
